[FIX] 설정 등록/수정 화면 버그 수정

+++ app/src/main/java/com/woowahantechcamp/account_book/data/entity/CategoryEntity.kt
@@ -4,5 +4,5 @@ data class CategoryEntity(
     val categoryId: Int,
     val type: Int,
     val title: String,
-    val color: Long
+    val color: Int
 )
